[build] Automatically add NullableAttributes.cs for netstandard2.0 (#1188)
`netstandard2.0` doesn't have "built-in" compiler support for
Nullable Reference Type attributes.  Thus when we enable nullable
reference types on `netstandard2.0` projects we are responsible for
providing the attributes ourselves.

Rather than manually repeat this info in each `.csproj` targeting
`netstandard2.0`, we can use `Directory.Build.targets` to automatically
add `NullableAttributes` in projects that need it.

Additionally, we can mark it as `Visible="false"` so the file doesn't
show up in Visual Studio Solution Explorer, just like the "real"
"built-in" attributes.