[docs] Reorganize public Documentation
Context: https://learn.microsoft.com/dotnet/android
Context: https://github.com/dotnet/docs-mobile
Context: https://dotnet.microsoft.com/en-us/platform/support/policy/xamarin

Support for Xamarin.Android ends on 2024-May-1.

Along with end of support, *documentation* for Xamarin.Android will
be *de-listed*: the content will still be there, but search will not
return results within it, and search engines will not find it.

Meaning that after May-1 there will be no documentation for error and
warning messages such as XA0000.

We now have a new documentation site for .NET for Android,
<https://learn.microsoft.com/dotnet/android>!

This new site takes content from
<https://github.com/dotnet/docs-mobile>, which isn't currently public,
but we are in the process of making it public.

Meanwhile, the `Documentation` directory was always intended to be
a source of *some* documentation, so that documentation for new
features could exist concurrent with the commit introducing the
feature.  This is not a workflow we wish to change.

To make synchronization easier, create a new
`Documentation/docs-mobile` directory which mirrors the contents of
[docs-mobile/docs/android][0], and migrate over our existing docs
into the new directory structure.

[0]: https://github.com/dotnet/docs-mobile/tree/main/docs/android