Re-enable marshal methods when ILStrip is enabled
The reason it broke before was that the
`EnableAndroidStripILAfterAOTFalse` runs in two steps - first it builds
the app, then it installs it (using the `Install` target).  During
the **build**, marshal methods classifier works as it should and,
consequently, marshal methods rewriter modifies the assemblies - removes
connector methods, generates wrapper methods etc etc.

`_GenerateJavaStubs`, the target which invokes marshal methods processing,
eventually creates a stamp file it then uses to decide whether all the
inputs are up to date with regards to their corresponding outputs.
However, during the **build** phase, at the end, `ILStrip` runs which
modifies assemblies **after** marshal methods processing is done.

This results in the `Install` target invoking the `_GenerateJavaStubs`
target again and MSBuild notices that the stamp file is older than some
of the inputs - assemblies modified by `ILStrip` after
`_GenerateJavaStubs` created the stamp file.  This causes the target to
run completely and, in effect, the whole marshal methods processing is
initiated again - but this time the connector methods and anything else
the classifier looks for aren't there, because they were previously
removed by the rewriter during the **build** phase.  This, in turn,
causes the classifier to decide that types need to be registered
dynamically, but they can't because the connector methods which are
required to create delegates are no longer there and we get a runtime
crash instead.

The solution is to update the `_GenerateJavaStubs` stamp file after
`ILStrip` is done, if it modified any assemblies.

One problem with that is that the path to the stamp file's directory is
different when `_GenerateJavaStubs` runs during the build phase, and
when AOT and `ILStrip` run.  In the former case, the stamp file is
created in `obj/${Configuration}/stamp` directory, in the latter case in
`obj/${Configuration}/android-ARCH/stamp` directory but **both**
locations are initialized in the same spot - at the top of
`Xamarin.Android.Common.targets` file.  In effect, after `ILStrip` runs,
we don't really know where `_GenerateJavaStubs` created its stamp file.
The workaround involves taking advantage of the fact that we know where
the stamp directories are in relation to each other.  It's a hack, but
if the relation is somehow broken, the
`EnableAndroidStripILAfterAOTFalse` test will break again.