Support Visual Studio "Build Acceleration"
Fixes #8581

The Visual Studio FastUpdate check flags the Resource Designer assembly as
newer than the output assembly (which it is). But it causes it to incorrectly
think the build is out of date.

```
WARNING: Potential build performance issue in 'Foo.csproj'.
The project does not appear up-to-date after a successful build:
Input ResolvedCompilationReference item 'obj\Debug\net8.0-android\_Microsoft.Android.Resource.Designer.dll' (2023-12-11 11:34:12.546) has been modified since the last successful build started
```

So lets remove it. We MUST remove the item from the @ReferencePathWithRefAssemblies
ItemGroup BEFORE the CollectResolvedCompilationReferencesDesignTime target runs.
This is because that target uses the Returns functionality. It turns out you cannot
modify the ItemGroup that is being returned via a Returns on a target.

Unfortunately there isn't a way to unit test this since the Fast Update check
ONLY runs in Visual Studio. The targets do not even exist on the command line.