[Java.Interop] Add `IJavaPeerable.JavaAs()` extension method (#1234)
Fixes: https://github.com/dotnet/java-interop/issues/10
Fixes: https://github.com/dotnet/android/issues/9038

Context: cf60cef071c81d3671e63c2ec7f318ed6cf91c76

Imagine the following Java type hierarchy:

	// Java
	public abstract class Drawable {
	    public static Drawable createFromStream(IntputStream is, String srcName) {…}
	    // …
	}
	public interface Animatable {
	    public void start();
	    // …
	}
	/* package */ class SomeAnimatableDrawable extends Drawable implements Animatable {
	    // …
	}

Further imagine that a call to `Drawable.createFromStream()` returns
an instance of `SomeAnimatableDrawable`.

What does the *binding* `Drawable.CreateFromStream()` return?

	// C#
	var drawable = Drawable.CreateFromStream(input, name);
	// What is the runtime type of `drawable`?

The binding `Drawable.CreateFromStream()` look at the runtime type
of the value returned, sees that it's of type `SomeAnimatableDrawable`,
and looks for an existing binding of that type.  If no such binding
is found -- which will be the case here, as `SomeAnimatableDrawable`
is package-private -- then we check the value's base class,
ad infinitum, until we hit a type that we *do* have a binding for
(or fail catastrophically if we can't find a binding for
`java.lang.Object`).  See also [`TypeManager.CreateInstance()`][0],
which is similar to the code within
`JniRuntime.JniValueManager.GetPeerConstructor()`.

Any interfaces implemented by Java value are not consulted, only
the base class hierarchy is consulted.

Consequently, the runtime type of `drawable` would be the
`Drawable` binding; however, as `Drawable` is an `abstract` type,
the runtime type will *actually* be `DrawableInvoker`
(see e.g. cf60cef0), akin to:

	// emitted by `generator`…
	internal class DrawableInvoker : Drawable {
	    // …
	}

Further imagine that we want to invoke `Animatable` methods on
`drawable`.  How do we do this?

This is where the [`.JavaCast<TResult>()` extension method][1] comes
in: we can use `.JavaCast<TResult>()` to perform a Java-side type
check for the desired type, which returns a value which can be used
to invoke methods on the specified type:

	var animatable = drawable.JavaCast<IAnimatable>();
	animatable.Start();

The problem with `.JavaCast<TResult>()` is that it always throws on
failure:

	var someOtherIface = drawable.JavaCast<ISomethingElse>();
	// throws some exception…

@mattleibow requests an "exception-free JavaCast overload" so that he
can *easily* use type-specific functionality *optionally*.

Add the following extension methods to `IJavaPeerable`:

	static partial class JavaPeerableExtensions {
	    public static TResult? JavaAs<TResult>(
	            this IJavaPeerable self);
	    public static bool TryJavaCast<TResult>(
	            this IJavaPeerable self,
	            out TResult? result);
	}

The `.JavaAs<TResult>()` extension method mirrors the C# `as`
operator, returning `null` if the the runtime type of `self`
is not implicitly convertible to the Java type corresponding to
`TResult`.  This makes it useful for one-off invocations:

	drawable.JavaAs<IAnimatable>()?.Start();

The `.TryJavaCast<TResult>()` extension method follows the
[`TryParse()` pattern][2], returning true if the type coercion
succeeds and the output `result` parameter is non-null, and false
otherwise.  This allows "nicely scoping" things within an `if`:

	if (drawable.TryJavaCast<IAnimatable>(out var animatable)) {
	    animatable.Start();
	    // …
	    animatable.Stop();
	}

[0]: https://github.com/dotnet/android/blob/06bb1dc6a292ef5618a3bb6ecca3ca869253ff2e/src/Mono.Android/Java.Interop/TypeManager.cs#L276-L291
[1]: https://github.com/dotnet/android/blob/06bb1dc6a292ef5618a3bb6ecca3ca869253ff2e/src/Mono.Android/Android.Runtime/Extensions.cs#L9-L17
[2]: https://learn.microsoft.com/dotnet/standard/design-guidelines/exceptions-and-performance#try-parse-pattern