[build] fix `ConfigureLocalWorkload` target
After 580166e6 was merged, the `ConfigureLocalWorkload` target was
no longer working as some of the structure of the `sdk-manifests`
folder has changed.

* We should run the `DeleteExtractedWorkloadPacks` target, as
  it will delete the `sdk-manifests/workloadset` folder.

Without doing this, the .NET SDK will *always* attempt to use the
34.99.0-preview.6.340 version in the "baseline" manifest.

* We also should put a versioned folder after each manifest directory.

So, for example:

    bin\Debug\lib\sdk-manifests\microsoft.net.sdk.android\ -> bin\Debug\lib\sdk-manifests\microsoft.net.sdk.android\35.0.0-preview.x.x\
    bin\Debug\lib\sdk-manifests\android.deps.workload\ -> bin\Debug\lib\sdk-manifests\android.deps.workload\0.0.1\

Now with a fresh checkout:

    dotnet msbuild -t:Prepare Xamarin.Android.sln
    .\dotnet-local.cmd build Xamarin.Android.sln
    .\dotnet-local.cmd build -t:ConfigureLocalWorkload build-tools\create-packs\Microsoft.Android.Sdk.proj

At this point, I can build app projects and the
`BuildBasicApplication` test passes locally.