[trimming] remove our usage of `[Preserve]`
Fixes: https://github.com/dotnet/android/issues/5197

`[Preserve]` still works in some form, as the `ApplyPreserveAttribute`
trimmer step will "mark" types it encounters. It unfortuantely, *does
not* work if an assembly is deemed to be fully removed, as the
`ApplyPreserveAttribute` will not be called for these types.

To better align with the future of .NET, we should remove *our* usage of
`[Preserve]`. The only failing result is `Java.Lang.Object.SetHandleOnDeserialized`
which can be preserved in `Mono.Android.xml` instead.

In a future PR, we can consider emitting warnings in customer code with
`[Preserve]` usage.