[ci] Disable CodeQL on CI/PR pipelines (#9128)
The conditions to disable CodeQL across different jobs and branches do
not appear be working as expected. The 'make prepare-update-mono' step
has been hanging intermittently on macOS, which appears to be due to
CodeQL.

Fix this by disabling CodeQL on this pipeline and only running it against
the nightly build.