Native runtime reorganization (#9113)
Context: https://github.com/dotnet/android/pull/9006

This commit prepares ground for dynamic native runtime linking at application build time by
moving code around and making it more modular, thus allowing the on-build linking of the 
runtime to work properly. 

In particular, this applies to the `p/invoke` handling code which is going to be different between the
pre-linked (and shipped) runtime and the one linked on demand during application build.  The difference
lies in the fact that on-demand build will scan the application for p/invoke usage (after trimming) and will
only include functions that are actually called from the managed land.   Results of the scan will be used to
generate LLVM IR code at build time that references the required functions, which will allow the native linker 
to remove the remaining ones. 

Changes implemented here can be committed to `main` and, possibly, released with .NET9
without changing how things currently work.